fix: extractorClient was replaced to /src
